// src/types.ts
export interface InventoryItem {
  id: string;
  name: string;
  category: 'Produce' | 'Meat & Poultry' | 'Dairy & Cheese' | 'Sauces & Condiments' | 'Bakery' | 'Bakery & Dough' | 'Dry Goods' | 'Beverages' | 'Packaging';
  currentStock: number;
  unit: 'kg' | 'portions' | 'liters' | 'units';
  minThreshold: number;
  costPerUnit: number;
  linkedMenuItems?: string[]; // item ids or modifier options
  isOut?: boolean;
}